Use o__Synechococcales instead of f__Synechococcaceae
This doesn't change the results at all (all of the features in
f__Synechococcaceae are also in o__Synechococcales). However, it looks
like there's disagreement over what family Prochlorococcus falls into
(but it's still in this order, at least) -- so this way we maintain
the analysis while respecting the new phylogeny.