redo analysis: use SILVA + diff Songbird params
the different Songbird params involve increasing the
num-random-test-examples by a small amount, and increasing the
batch-size by a large amount. from what i've seen this results in
a cleaner-looking "loss" diagnostic plot, which is why i took time
today to tinker around with this... granted, i don't think this
impacts results too much. (it does change around where the topmost
shewanella is ranked slightly, but it's near the top pretty much
however you do this so it shouldn't impact the paper much or at all)

use of SILVA for taxonomy here closes #5.