RM negative_control_stats
use of hack in which we only classify certain features with taxonomy
means that it won't be super useful for this repo anymore.
furthermore, use of katharoseq/protocols for mitigating well-to-well
contamination, etc. means that we shouldn't expect contamination to
be a big deal here anyway.

i put the script over on my github gists page, so it's there for
future reference